feat: add armor bonus display in unit sheet
